Update Find_New_Repeats.py
Preliminary to the following - I may need to consider it a bit more, but I'm not entirely sure we can fully take advantage of Karp-Rabin in this version of the code. I will explain this tomorrow as I'm not entirely sure at the moment.

A larger note - you should separate the hashing process into its own functions so that we can improve the readability of the main process
in this code - as it stands now, at least as someone coming in and reading this code, it takes me some time to differentiate the lines that are hashing and rehashing vs actually collecting pattern information from the text.

Separating these will allow us to mutually exclude the two functions from each other when we do testing, and thus we will be able to test and debug them separately. I've added headers for hash and rehash to the top of this file, but you'll need to implement them.

To get you motivated a bit, identify the lines where you hash, identify the lines where you rehash, and finally replace those lines with the appropriate function calls (hint - you may not need to write rehash separately from hash, as their is never overlap between the curr and prev chunks and thus you will be rehashing a number of characters equal to the size of the pattern with every hash, which is, by our scheme, equivalent to just calling hash on those characters).

Additionally, consider how many things that you actually need to hash per loop - if we have a previous and a current chunk, and then proceed to the next iteration, what will our new previous block be (big hint: you've already computed the hash for this chunk (bigger hint: the old current chunk becomes the new previous chunk))?

I also ran through the code and modified a few of the simpler lines: 
Modified the following lines:
- 15: Changed class name to STR to follow class naming style
- 66: Flipped the terms in the 'and' expression as the order has non-trivial effect on efficiency
- Throughout: Removed a few lines and moved a few lines around, mostly concerning the current and previous text to ensure that they are set properly

Added comments on the following lines:
- STR Definition: Added a few explanatory comments
- 64: A note about how to interpret i's value
- Added a lot of hints and other notes in the "patterns_search" code - I recommend you read over these